Add editClient script
In the spirit of createClient and startClient scripts, editClient allows one to
evaluate statements in a client image(s).

Rationale for creating this script - after creating multiple client images
(using prefix), one may be inclined to evaluate image-specific code in each of
the images. This could be - configuring seaside options that are specific to
each image, or change error handler to help debug a problem...